+++ src/lib/app-provider/hooks/index.ts
@@ -19,4 +19,5 @@ export * from "./useIsConnected";
 export * from "./useChainConfigs";
 export * from "./useAllowCustomNetworks";
 export * from "./useSignAndBroadcast";
+export * from "./useSignAndBroadcastEvm";
 export * from "./useSimulateFee";