+++ src/lib/services/tx/jsonRpc.ts
@@ -1,21 +1,23 @@
 import { z } from "zod";
 
+import { TransactionRequest } from "ethers";
+import { parseWithError } from "lib/utils";
 import type { JsonRpcParams } from "../evm/jsonRpc";
 import { requestBatchJsonRpc, requestJsonRpc } from "../evm/jsonRpc";
 import {
   zEvmTxHashByCosmosTxHashJsonRpc,
   zEvmTxHashesByCosmosTxHashesJsonRpc,
+  zSimulateFeeEvm,
   zTxDataJsonRpc,
   zTxsDataJsonRpc,
 } from "../types";
-import { parseWithError } from "lib/utils";
 
 export const getEvmTxHashByCosmosTxHash = async (
   endpoint: string,
   cosmosTxHash: string
 ) =>
   requestJsonRpc(endpoint, "cosmos_txHashByCosmosTxHash", [cosmosTxHash]).then(
-    ({ result }) => parseWithError(zEvmTxHashByCosmosTxHashJsonRpc, result)
+    (result) => parseWithError(zEvmTxHashByCosmosTxHashJsonRpc, result)
   );
 
 export const getEvmTxHashesByCosmosTxHashes = async (
@@ -81,5 +83,20 @@ export const getCosmosTxHashByEvmTxHash = (
   evmTxHash: string
 ) =>
   requestJsonRpc(endpoint, "cosmos_cosmosTxHashByTxHash", [evmTxHash]).then(
-    ({ result }) => parseWithError(z.string(), result)
+    (result) => parseWithError(z.string(), result)
   );
+
+export const getSimulateFeeEvm = (
+  endpoint: string,
+  request: TransactionRequest
+) =>
+  requestBatchJsonRpc(endpoint, [
+    {
+      method: "eth_gasPrice",
+      params: [],
+    },
+    {
+      method: "eth_estimateGas",
+      params: [request],
+    },
+  ]).then((results) => parseWithError(zSimulateFeeEvm, results));