+++ src/lib/utils/mergeRefs.ts
@@ -10,7 +10,7 @@ export const mergeRefs = <T = unknown>(
     refs.forEach((ref) => {
       if (typeof ref === "function") {
         ref(value);
-      } else if (ref != null) {
+      } else if (ref !== null) {
         (ref as React.MutableRefObject<Nullable<T>>).current = value;
       }
     });